python-server-plugin/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "deephaven-rpc-pickle-server-plugin"
version = "0.0.1"
dependencies = ["deephaven-plugin>=0.5.0", "deephaven-core>=0.27.1"]

[project.entry-points."deephaven.plugin"]
registration_cls = "dhexample.rpc_pickle_server.plugin:RpcPicklePluginRegistration"
